drop_acc {rdrop2}	R Documentation

Get information about current Dropbox account.

Description

Fields returned will vary by account;

Usage

drop_acc(dtoken = get_dropbox_token())

Arguments

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

Value

Nested list with elements account_id, name (list), email, email_verified, disabled, locale, referral_link, is_paired, account_type (list).

If available, may also return profile_photo_url, country, team (list), team_member_id.

References

API documentation

Examples

## Not run: 

  acc_info <- drop_acc()

  # extract display name
  acc_info$name$display_name

## End(Not run)

[Package rdrop2 version 0.8.1 Index]