drop_copy {rdrop2}	R Documentation

Copies a file or folder to a new location.

Description

Copies a file or folder to a new location.

Usage

drop_copy(from_path = NULL, to_path = NULL, allow_shared_folder = FALSE,
  autorename = FALSE, allow_ownership_transfer = FALSE, verbose = FALSE,
  dtoken = get_dropbox_token())

Arguments

from_path
Source file or folder

to_path
destination file or folder

allow_shared_folder
If TRUE, copy will copy contents in shared folder

autorename
If there's a conflict, have the Dropbox server try to autorename the file to avoid the conflict.

allow_ownership_transfer
Allow moves by owner even if it would result in an ownership transfer for the content being moved. This does not apply to copies. The default for this field is False.

verbose
By default verbose output is FALSE. Set to TRUE if you need to troubleshoot any output or grab additional parameters.

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

References

API documentation

Examples

## Not run: 
write.csv(mtcars, file = "mt.csv")
drop_upload("mt.csv")
drop_create("drop_test2")
drop_copy("mt.csv", "drop_test2/mt2.csv")

## End(Not run)

[Package rdrop2 version 0.8.1 Index]