drop_delete {rdrop2}	R Documentation

Deletes a file or folder.

Description

Deletes a file or folder.

Usage

drop_delete(path = NULL, verbose = FALSE, dtoken = get_dropbox_token())

Arguments

path
This is required The path to the new folder to create relative to root.

verbose
By default verbose output is FALSE. Set to TRUE if you need to troubleshoot any output or grab additional parameters.

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

References

API documentation

[Package rdrop2 version 0.8.1 Index]