drop_get {rdrop2}	R Documentation

Downloads a file from Dropbox

Description

Downloads a file from Dropbox

Usage

drop_get(path = NULL, local_file = NULL, overwrite = FALSE,
  verbose = FALSE, progress = FALSE, dtoken = get_dropbox_token())

Arguments

path
This is required The path to the new folder to create relative to root.

local_file
The name of the local copy. Leave this blank if you're fine with the original name.

overwrite
Default is FALSE but can be set to TRUE.

verbose
By default verbose output is FALSE. Set to TRUE if you need to troubleshoot any output or grab additional parameters.

progress
Progress bars are turned off by default. Set to TRUE ot turn this on. Progress is only reported when file sizes are known. Otherwise just bytes downloaded.

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

Examples

## Not run: 
  drop_get(path = 'dataset.zip', local_file = "~/Desktop")
  # To overwrite the existing file
  drop_get(path = 'dataset.zip', overwrite = TRUE)

## End(Not run)

[Package rdrop2 version 0.8.1 Index]