drop_get_metadata {rdrop2}	R Documentation

Retrieve metadata for a file or folder.

Description

Details vary by input and args.

Usage

drop_get_metadata(path, include_media_info = FALSE, include_deleted = FALSE,
  include_has_explicit_shared_members = FALSE, dtoken = get_dropbox_token())

Arguments

path
Path to a file or folder on Dropbox. Can also be an ID ("id:...") or revision ("rev:...").

include_media_info
If TRUE, additional metadata for photo or video is returns. Defaults to FALSE.

include_deleted
If TRUE, metadata will be returned for a deleted file, otherwise error. Defaults to FALSE.

include_has_explicit_shared_members
If TRUE, the results will include a flag for each file indicating whether or not that file has any explicit members. Defaults to FALSE.

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

Value

possibly-nested list of all available metadata for specified file/folder/id/revision.

References

API Documentation

[Package rdrop2 version 0.8.1 Index]