drop_history {rdrop2}	R Documentation

Obtains metadata for all available revisions of a file, including the current revision.

Description

Does not include deleted revisions.

Usage

drop_history(path, limit = 10, dtoken = get_dropbox_token())

Arguments

path
path to a file in dropbox.

limit
maximum number of revisions to return; defaults to 10.

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

Value

tbl_df of metadata, one row per revision.

Examples

## Not run: 
  write.csv(iris, file = "iris.csv")
  drop_upload("iris.csv")
  write.csv(iris[iris$Species == "setosa", ], file = "iris.csv")
  drop_upload("iris.csv")
  drop_history("iris.csv")

## End(Not run)

[Package rdrop2 version 0.8.1 Index]