drop_list_shared_links {rdrop2}	R Documentation

List all shared links

Description

This function returns a list of all links that are currently being shared

Usage

drop_list_shared_links(verbose = TRUE, dtoken = get_dropbox_token())

Arguments

verbose
Print verbose output

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

References

API documentation

Examples

## Not run: 
drop_list_shared_links()

## End(Not run)

[Package rdrop2 version 0.8.1 Index]