drop_read_csv {rdrop2}	R Documentation

drop_read_csv

Description

A lightweight wrapper around read.csv to read csv files from Dropbox into memory

Usage

drop_read_csv(file, dest = tempdir(), dtoken = get_dropbox_token(), ...)

Arguments

file
Name of file with full path relative to Dropbox root

dest
A temporary directory where a csv file is downloaded before being read into memory

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

...
Additional arguments into read.csv

Examples

## Not run: 
write.csv(iris, file = "iris.csv")
drop_upload("iris.csv")
# Now let's read this back into an R session
new_iris <- drop_read_csv("iris.csv")

## End(Not run)

[Package rdrop2 version 0.8.1 Index]