Returns metadata for all files and folders whose filename contains the given search string as a substring.

Returns metadata for all files and folders whose filename contains the given search string as a substring.

drop_search(query, path = "", start = 0, max_results = 100,
  mode = "filename", dtoken = get_dropbox_token())

Arguments

query
The search string. This string is split (on spaces) into individual words. Files and folders will be returned if they contain all words in the search string.

path
This is required The path to the new folder to create relative to root.

start
The starting index within the search results (used for paging). The default for this field is 0

max_results
The maximum number of search results to return. The default for this field is 100.

mode
Mode can take the option of filename, filename_and_content, or search deleted files with deleted_filename

dtoken
The Dropbox token generated by drop_auth. rdrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this in drop_auth by pointing to a different location where your credentials are stored.

Examples

# NOT RUN {
# If you know me, you know why this query exists
drop_search('gif') %>% select(path, is_dir, mime_type)
# }